app now working with three list itemsafter fixing some confusing errors

+++ app/src/test/java/example/codeclan/com/magazines/MagazineTest.java
@@ -25,11 +25,11 @@ public class MagazineTest {
 
     @Test
     public void getRankingTest(){
-        assertEquals(1, magazine.getRanking());
+        assertEquals((Integer)1, magazine.getRanking());
     }
 
     @Test
     public void getYearTest(){
-        assertEquals(2016, magazine.getYear());
+        assertEquals((Integer)2016, magazine.getYear());
     }
 }
